pyproject.toml
[project]
name = "comfyui-dci"
description = "A comprehensive ComfyUI extension for creating, previewing, and analyzing DCI (DSG Combined Icons) format files. Supports multi-state icons, multi-tone variants, scaling factors, and advanced metadata analysis with complete Chinese localization. Also provides universal binary file loading/saving and base64 encoding/decoding nodes for general file processing workflows."
version = "1.0.0"
license = { file = "LICENSE" }
dependencies = [
    "Pillow>=8.0.0",
    "numpy>=1.19.0"
]
readme = "README.md"
requires-python = ">=3.8"
keywords = ["comfyui", "dci", "icons", "image-processing", "desktop-icons", "binary-files", "file-loader", "file-saver", "base64", "encoding", "decoding", "file-processing", "workflow-tools"]
authors = [
    { name = "ComfyUI DCI Extension Team" }
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Archiving",
    "Topic :: Utilities",
    "Operating System :: OS Independent"
]

[project.urls]
Repository = "https://github.com/zccrs/comfyui-dci"
Homepage = "https://github.com/zccrs/comfyui-dci"
Documentation = "https://github.com/zccrs/comfyui-dci/blob/master/README.md"
"Bug Reports" = "https://github.com/zccrs/comfyui-dci/issues"

[tool.comfy]
PublisherId = "zccrs"
DisplayName = "DCI Image Export Extension"
Icon = "https://raw.githubusercontent.com/zccrs/comfyui-dci/master/resources/icon.svg"

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["py*", "locales*", "resources*"]
exclude = ["tests*", "test_env*", "tools*", "examples*", "venv*"]

[tool.setuptools.package-data]
"*" = ["*.json", "*.png", "*.jpg", "*.svg", "*.md", "*.txt"]

# Test configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning"
]

[tool.coverage.run]
source = ["py"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/venv/*",
    "*/__pycache__/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:"
]
